increasing a lot of size limits in prep for doom 1/2, and shuffling some data around to make it happen. big issue is going to be texturecolumnofs and texturecolumnlumps which are increasing from 21552 bytes each to 80480 bytes each (not implemented yet). will need to address in later commit with new allocation/pagination strategy

+++ p_floor.c
@@ -413,7 +413,7 @@ EV_DoFloor
 		  case raiseToTexture: {
 			  short_height_t minsize = MAXSHORT;
 			  int16_t sidenum;
-			  uint8_t sidebottomtexture;
+			  uint16_t sidebottomtexture;
 				
 			  floor->direction = 1;
 			  floor->secnum = secnum;
